Add UMN parameter group, add/move parameters, bump xml minor version.
Added UMN parameter group with parameters defined in umn_params.c.
Updated prior EKF2_UMN_CONTROL param to be part of UMN group.
Added new parameter UMN_GPS_ON to control availability of GPS.
Bumped the minor version for the parameters xml generator. This is based
on the following reference:
http://discuss.px4.io/t/changing-parameters-meta-in-qgroundcontrol/594/5

Note that the new firmware must be uploaded via QGC for the new parameters
and group to be visible.  This is also true for seeing the changes in SITL.