Expand UMN message with ekf2 gyro bias and track GPS status.
EKF2 was updated to supply gyro bias estimates to the UMN message.
This is done only if the UMN_GPS_ON flag is `true`.  The UMN app
is also set to only update the local GPS uORB copy if the UMN_GPS_ON
flag is `true`.  Otherwise, the stale gps message `satellites_used`
will be set to 0 and the stale data will be passed to the UMN
algorithm.
The UMN_GPS_ON flag is now logged with the umn output messages.