Fix bug where relevant QGC parameter states were not updating.
Two states were added to the UMN message to reflect the QGC
parameters.  These were properly updated for logging but not
for usage in the umn_algorithm function.  This is now changed
so that the values are correct in both scenarios.  Also, the variable
types have been updated from bool to int so that exact values can be
observed.